Fix workspace restore policy review gates
